Added disable gridlines feature to chart

+++ DbD_Pingz/PacketStats.cs
@@ -40,7 +40,7 @@ namespace DbD_Pingz
         {
             if(total > 0)
             {
-                PacketLoss = (float) ((float)dismissed / (float)total);
+                PacketLoss = (float) dismissed / total;
             }
         }
